requirements.txt
openai-whisper>=20231117
ffmpeg-python>=0.2.0
torch>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
click>=8.1.0
